# docs/swagger.yaml
basePath: /
definitions:
...
  admin.createAgesRequest:
    properties:
      poster:
        $ref: '#/definitions/multipart.FileHeader'
      title:
        type: string
    type: object
...
  admin.createMovieRequest:
    properties:
      ages:
        items:
          type: integer
        type: array
      categories:
        items:
          type: integer
        type: array
      description:
        type: string
      director:
        type: string
      genres:
        items:
          type: integer
        type: array
      keywords:
        items:
          type: string
        type: array
      movieTypeId:
        type: integer
      producer:
        type: string
      releaseYear:
        type: integer
      runtime:
        type: integer
      title:
        type: string
    type: object
...
  admin.createRoleRequest:
    properties:
      can_edit_ages:
        type: boolean
      can_edit_categories:
        type: boolean
      can_edit_genres:
        type: boolean
      can_edit_projects:
        type: boolean
      can_edit_roles:
        type: boolean
      can_edit_users:
        type: boolean
      name:
        type: string
    required:
    - name
    type: object
  admin.userResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  gin.H:
    additionalProperties: {}
    type: object
  models.Ages:
    properties:
      id:
        type: integer
      posterUrl:
        type: string
      title:
        type: string
    type: object
  models.ApiError:
    properties:
      error:
        type: string
    type: object
  models.Category:
    properties:
      id:
        type: integer
      title:
        type: string
    type: object
  models.Episode:
    properties:
      id:
        type: integer
      number:
        description: Номер серии
        type: integer
      seasonID:
        description: ID сезона, к которому относится эпизод
        type: integer
      videoURL:
        description: Ссылка на видео
        type: string
    type: object
  models.Genre:
    properties:
      id:
        type: integer
      posterUrl:
        type: string
      title:
        type: string
    type: object
  models.Movie:
    properties:
      ages:
        items:
          $ref: '#/definitions/models.Ages'
        type: array
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      cover:
        type: string
      description:
        type: string
      director:
        type: string
      genres:
        items:
          $ref: '#/definitions/models.Genre'
        type: array
      id:
        type: integer
      keyWords:
        items:
          type: string
        type: array
      movieType:
        type: string
      movieTypeId:
        type: integer
      producer:
        type: string
      releaseYear:
        type: integer
      runtime:
        type: integer
      screenshots:
        items:
          type: string
        type: array
      seasons:
        items:
          $ref: '#/definitions/models.Season'
        type: array
      title:
        type: string
    type: object
...
  models.Role:
    properties:
      canEditAges:
        type: boolean
      canEditCategories:
        type: boolean
      canEditGenres:
        type: boolean
      canEditProjects:
        type: boolean
      canEditRoles:
        type: boolean
      canEditUsers:
        type: boolean
      id:
        type: integer
      name:
        type: string
    type: object
  models.Season:
    properties:
      episodes:
        description: Связь один ко многим
        items:
          $ref: '#/definitions/models.Episode'
        type: array
      id:
        type: integer
      movieID:
        description: ID фильма, к которому относится сезон
        type: integer
      number:
        description: Номер сезона
        type: integer
    type: object
...
  multipart.FileHeader:
    properties:
      filename:
        type: string
      header:
        $ref: '#/definitions/textproto.MIMEHeader'
      size:
        type: integer
    type: object
...
  textproto.MIMEHeader:
    additionalProperties:
      items:
        type: string
      type: array
    type: object
host: localhost:8081
info:
  contact:
    email: bexultanovmadiyar@gmail.com
    name: Madiyar Bexultanov
    url: https://www.linkedin.com/in/madiyar-bexultanov-b21902258/
  description: Personal online platform providing information about media content
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Ozinshe Production
  version: "1.0"
paths:
  /admin/ages:
    get:
      description: Retrieves a list of all ages
      produces:
      - application/json
      responses:
        "200":
          description: A list of ages
          schema:
            items:
              $ref: '#/definitions/models.Ages'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get all ages
      tags:
      - ages
    post:
      consumes:
      - application/json
      description: Creates a new age entry
      parameters:
      - description: Create age request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.createAgesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ID of the created age
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Failed to save poster or create age
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Create new age
      tags:
      - ages
...
  /admin/genres:
    get:
      consumes:
      - application/json
      description: Retrieve all genres from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Genre'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get all genres
      tags:
      - genres
    post:
      consumes:
      - multipart/form-data
      description: Create a new genre and upload a poster
      parameters:
      - description: Title of the genre
        in: formData
        name: title
        required: true
        type: string
      - description: Poster image
        in: formData
        name: poster
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: id of the created genre
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Create a new genre
      tags:
      - genres
...
  /admin/movies:
    get:
      consumes:
      - application/json
      description: Get a list of all movies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Movie'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get all movies
      tags:
      - Movies
    post:
      consumes:
      - application/json
      description: Create a new movie with the specified information
      parameters:
      - description: Movie information
        in: body
        name: movie
        required: true
        schema:
          $ref: '#/definitions/admin.createMovieRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Movie ID
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Create a new movie
      tags:
      - Movies
...
  /admin/roles:
    get:
      description: Get a list of all roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Role'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get all roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Create a new role with specified permissions
      parameters:
      - description: Role data
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/admin.createRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ID of the created role
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Create a new role
      tags:
      - Roles
...
  /admin/users:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/admin.userResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get all users
      tags:
      - Users
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"